# docs/swagger.yaml
basePath: /
definitions:
...
  internal_adapters_primary_api_user.changePasswordProfileReq:
    properties:
      current_password:
        type: string
      id:
        type: string
      new_password:
        type: string
    type: object
  internal_adapters_primary_api_user.deleteAccountReq:
    properties:
      id:
        type: string
      session_id:
        type: string
    type: object
  internal_adapters_primary_api_user.getUserProfileReq:
    properties:
      id:
        type: string
    type: object
  internal_adapters_primary_api_user.getUserProfileResp:
    properties:
      email:
        type: string
      id:
        type: string
      is_admin:
        type: boolean
      name:
        type: string
    type: object
  internal_adapters_primary_api_user.listUsersResp:
    properties:
      users:
        items:
          $ref: '#/definitions/internal_adapters_primary_api_user.userInfo'
        type: array
    type: object
  internal_adapters_primary_api_user.loginUserReq:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  internal_adapters_primary_api_user.loginUserResp:
    properties:
      access_token:
        type: string
      access_token_expires_at:
        type: string
      email:
        type: string
      is_admin:
        type: boolean
      name:
        type: string
      refresh_token:
        type: string
      refresh_token_expires_at:
        type: string
      session_id:
        type: string
    type: object
  internal_adapters_primary_api_user.registerUserRequest:
    properties:
      email:
        type: string
      is_admin:
        type: boolean
      name:
        type: string
      password:
        type: string
    type: object
  internal_adapters_primary_api_user.registerUserResp:
    properties:
      id:
        type: string
    type: object
  internal_adapters_primary_api_user.renewAccessTokenReq:
    properties:
      refresh_token:
        type: string
    type: object
  internal_adapters_primary_api_user.renewAccessTokenResp:
    properties:
      access_token:
        type: string
      access_token_expires_at:
        type: string
    type: object
  internal_adapters_primary_api_user.updateUserProfileReq:
    properties:
      email:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
  internal_adapters_primary_api_user.userInfo:
    properties:
      email:
        type: string
      id:
        type: string
      is_admin:
        type: boolean
      name:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
  description: |-
    A RESTful e-commerce API built with Go using hexagonal architecture.
    Features: User authentication (JWT), Product management, Order processing, Address management.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: E-Commerce API
  version: "1.0"
paths:
...
  /admin/users:
    get:
      consumes:
      - application/json
      description: Get a list of all users (admin only)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_adapters_primary_api_user.listUsersResp'
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: List all users (Admin)
      tags:
      - Admin
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user and return tokens
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_adapters_primary_api_user.loginUserReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_adapters_primary_api_user.loginUserResp'
        "400":
          description: Invalid request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
      summary: Login user
      tags:
      - Auth
  /auth/logout:
    post:
      consumes:
      - application/json
      description: Invalidate user session
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Logout user
      tags:
      - Auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Create a new user account
      parameters:
      - description: User registration data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_adapters_primary_api_user.registerUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_adapters_primary_api_user.registerUserResp'
        "400":
          description: Invalid request
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Register a new user
      tags:
      - Auth
  /auth/renew:
    post:
      consumes:
      - application/json
      description: Get a new access token using refresh token
      parameters:
      - description: Refresh token
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_adapters_primary_api_user.renewAccessTokenReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_adapters_primary_api_user.renewAccessTokenResp'
        "400":
          description: Invalid request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
      summary: Renew access token
      tags:
      - Auth
...
  /users/{id}:
    delete:
      consumes:
      - application/json
      description: Delete user account
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Account deletion data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_adapters_primary_api_user.deleteAccountReq'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "400":
          description: Invalid request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete account
      tags:
      - Users
    get:
      consumes:
      - application/json
      description: Get user profile by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: User ID
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_adapters_primary_api_user.getUserProfileReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_adapters_primary_api_user.getUserProfileResp'
        "400":
          description: Invalid request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get user profile
      tags:
      - Users
    put:
      consumes:
      - application/json
      description: Update user profile information
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated profile data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_adapters_primary_api_user.updateUserProfileReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Invalid request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Update user profile
      tags:
      - Users
  /users/{id}/password:
    put:
      consumes:
      - application/json
      description: Change user password
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Password change data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_adapters_primary_api_user.changePasswordProfileReq'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "400":
          description: Invalid request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Change password
      tags:
      - Users
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"